feat: move healthcheck to compose (#7)
Adding HEALTHCHECK in Dockerfile is activated by default. The
healthcheck at the scheduler level could be different for compose and
kubernetes and moving the healthcheck into a script is a better
trade-off. Plus, on local dev env it could slowdown the deployment since
it wait for the first interval + start period.
- [x] add healthcheck.sh script
- [x] move healthcheck to compose
- [x] update README
- [x] update CI